requirements.txt
langchain==0.0.133
openai==0.27.0
faiss-cpu==1.7.3
tiktoken==0.3.3
supabase==1.0.3
streamlit-modal==0.1.0
streamlit-analytics